tests/requirements-test.txt
# Testing dependencies
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0